feat: Add NameAndOtherInputs and UploadAssetsModal components for enhanced workout management
