+++ apps/frontend/app/routes/_dashboard.fitness.$action.tsx
@@ -1,44 +1,20 @@
 import { useAutoAnimate } from "@formkit/auto-animate/react";
 import {
-	ActionIcon,
-	Anchor,
-	Avatar,
-	Box,
 	Button,
 	Container,
 	Divider,
-	FileButton,
 	Group,
-	Modal,
-	NumberInput,
 	SimpleGrid,
 	Skeleton,
 	Stack,
-	Text,
-	TextInput,
-	Textarea,
 } from "@mantine/core";
-import { useDebouncedState, useDidUpdate, useDisclosure } from "@mantine/hooks";
+import { useDisclosure } from "@mantine/hooks";
 import { notifications } from "@mantine/notifications";
 import {
 	CreateOrUpdateUserWorkoutDocument,
 	CreateOrUpdateUserWorkoutTemplateDocument,
-	GetPresignedS3UrlDocument,
 } from "@ryot/generated/graphql/backend/graphql";
-import {
-	changeCase,
-	isNumber,
-	isString,
-	parseParameters,
-	sum,
-} from "@ryot/ts-utils";
-import {
-	IconCamera,
-	IconLibraryPhoto,
-	IconTrash,
-	IconVideo,
-} from "@tabler/icons-react";
-import { useQuery } from "@tanstack/react-query";
+import { changeCase, isString, parseParameters, sum } from "@ryot/ts-utils";
 import clsx from "clsx";
 import { Howl } from "howler";
 import { produce } from "immer";
@@ -47,7 +23,6 @@ import { useState } from "react";
 import { Link, useLoaderData, useNavigate } from "react-router";
 import { ClientOnly } from "remix-utils/client-only";
 import { $path } from "safe-routes";
-import invariant from "tiny-invariant";
 import { match } from "ts-pattern";
 import { useInterval } from "usehooks-ts";
 import { z } from "zod";
@@ -58,6 +33,10 @@ import {
 	getProgressOfExercise,
 	usePerformTasksAfterSetConfirmed,
 } from "~/components/fitness.action/hooks";
+import {
+	NameAndOtherInputs,
+	UploadAssetsModal,
+} from "~/components/fitness.action/miscellaneous";
 import { ReorderDrawer } from "~/components/fitness.action/reorder";
 import {
 	RestTimer,
@@ -70,9 +49,7 @@ import { deleteUploadedAsset } from "~/components/fitness.action/utils";
 import {
 	FitnessAction,
 	FitnessEntity,
-	PRO_REQUIRED_MESSAGE,
 	clientGqlService,
-	clientSideFileUpload,
 	dayjsLib,
 	openConfirmationModal,
 	postMessageToServiceWorker,
@@ -82,13 +59,11 @@ import {
 } from "~/lib/common";
 import {
 	useApplicationEvents,
-	useCoreDetails,
 	useUserPreferences,
 	useUserUnitSystem,
 } from "~/lib/hooks";
 import {
 	currentWorkoutToCreateWorkoutInput,
-	getExerciseDetailsQuery,
 	useCurrentWorkout,
 	useCurrentWorkoutTimerAtom,
 	useMeasurementsDrawerOpen,
@@ -571,312 +546,3 @@ export default function Page() {
 		</Container>
 	);
 }
-
-const NameAndOtherInputs = (props: {
-	isCreatingTemplate: boolean;
-	openAssetsModal: () => void;
-}) => {
-	const userPreferences = useUserPreferences();
-	const [currentWorkout, setCurrentWorkout] = useCurrentWorkout();
-	invariant(currentWorkout);
-
-	const [name, setName] = useDebouncedState(currentWorkout.name, 500);
-	const [comment, setComment] = useDebouncedState(currentWorkout.comment, 500);
-	const [isCaloriesBurntModalOpen, setIsCaloriesBurntModalOpen] =
-		useState(false);
-	const [caloriesBurnt, setCaloriesBurnt] = useDebouncedState(
-		currentWorkout.caloriesBurnt,
-		500,
-	);
-	const workoutHasImages = currentWorkout.images.length > 0;
-
-	useDidUpdate(() => {
-		if (name)
-			setCurrentWorkout(
-				produce(currentWorkout, (draft) => {
-					draft.name = name;
-				}),
-			);
-	}, [name]);
-
-	useDidUpdate(() => {
-		setCurrentWorkout(
-			produce(currentWorkout, (draft) => {
-				draft.comment = comment || undefined;
-			}),
-		);
-	}, [comment]);
-
-	useDidUpdate(() => {
-		setCurrentWorkout(
-			produce(currentWorkout, (draft) => {
-				draft.caloriesBurnt = caloriesBurnt;
-			}),
-		);
-	}, [caloriesBurnt]);
-
-	return (
-		<>
-			<Modal
-				title="Additional details"
-				opened={isCaloriesBurntModalOpen}
-				onClose={() => setIsCaloriesBurntModalOpen(false)}
-			>
-				<Stack gap="xs">
-					<NumberInput
-						size="sm"
-						value={currentWorkout.caloriesBurnt}
-						label={`Energy burnt in ${userPreferences.fitness.logging.caloriesBurntUnit}`}
-						onChange={(e) => setCaloriesBurnt(isNumber(e) ? e : undefined)}
-					/>
-					<Textarea
-						size="sm"
-						minRows={2}
-						label="Comments"
-						defaultValue={comment}
-						placeholder="Your thoughts about this workout"
-						onChange={(e) => setComment(e.currentTarget.value)}
-					/>
-				</Stack>
-			</Modal>
-			<TextInput
-				size="sm"
-				defaultValue={name}
-				placeholder="A name for your workout"
-				styles={{ label: { width: "100%" } }}
-				onChange={(e) => setName(e.currentTarget.value)}
-				rightSection={
-					<ActionIcon
-						onClick={props.openAssetsModal}
-						variant={workoutHasImages ? "outline" : undefined}
-					>
-						<IconCamera size={30} />
-					</ActionIcon>
-				}
-				label={
-					<Group justify="space-between" mr="xs">
-						<Text size="sm">Name</Text>
-						{!props.isCreatingTemplate ? (
-							<Anchor
-								size="xs"
-								onClick={() => setIsCaloriesBurntModalOpen(true)}
-							>
-								More Information
-							</Anchor>
-						) : null}
-					</Group>
-				}
-			/>
-		</>
-	);
-};
-
-const AssetDisplay = (props: {
-	s3Key: string;
-	type: "video" | "image";
-	removeAsset: () => void;
-}) => {
-	const srcUrlQuery = useQuery({
-		queryKey: queryFactory.miscellaneous.presignedS3Url(props.s3Key).queryKey,
-		queryFn: () =>
-			clientGqlService
-				.request(GetPresignedS3UrlDocument, { key: props.s3Key })
-				.then((v) => v.getPresignedS3Url),
-	});
-
-	return (
-		<Box pos="relative">
-			{props.type === "video" ? (
-				<Link to={srcUrlQuery.data ?? ""} target="_blank">
-					<Avatar size="lg" name="Video" />
-				</Link>
-			) : (
-				<Avatar src={srcUrlQuery.data} size="lg" />
-			)}
-			<ActionIcon
-				top={0}
-				size="xs"
-				left={-12}
-				color="red"
-				pos="absolute"
-				onClick={() => {
-					openConfirmationModal(
-						"Are you sure you want to remove this video?",
-						() => props.removeAsset(),
-					);
-				}}
-			>
-				<IconTrash />
-			</ActionIcon>
-		</Box>
-	);
-};
-
-const UploadAssetsModal = (props: {
-	closeModal: () => void;
-	modalOpenedBy: string | null | undefined;
-}) => {
-	const coreDetails = useCoreDetails();
-	const fileUploadAllowed = coreDetails.fileStorageEnabled;
-	const [currentWorkout, setCurrentWorkout] = useCurrentWorkout();
-	const [isFileUploading, setIsFileUploading] = useState(false);
-
-	if (!currentWorkout) return null;
-
-	const afterFileSelected = async (
-		file: File | null,
-		type: "image" | "video",
-	) => {
-		if (props.modalOpenedBy === null && !coreDetails.isServerKeyValidated) {
-			notifications.show({
-				color: "red",
-				message: PRO_REQUIRED_MESSAGE,
-			});
-			return;
-		}
-		if (!file) return;
-		setIsFileUploading(true);
-		try {
-			const key = await clientSideFileUpload(file, "workouts");
-			setCurrentWorkout(
-				produce(currentWorkout, (draft) => {
-					if (type === "image") {
-						if (exercise) draft.exercises[exerciseIdx].images.push(key);
-						else draft.images.push(key);
-					} else {
-						if (exercise) draft.exercises[exerciseIdx].videos.push(key);
-						else draft.videos.push(key);
-					}
-				}),
-			);
-		} catch {
-			notifications.show({
-				color: "red",
-				message: `Error while uploading ${type}`,
-			});
-		} finally {
-			setIsFileUploading(false);
-		}
-	};
-
-	const exerciseIdx = currentWorkout.exercises.findIndex(
-		(e) => e.identifier === props.modalOpenedBy,
-	);
-	const exercise =
-		exerciseIdx !== -1 ? currentWorkout.exercises[exerciseIdx] : null;
-
-	const { data: exerciseDetails } = useQuery({
-		...getExerciseDetailsQuery(exercise?.exerciseId || ""),
-		enabled: exercise !== null,
-	});
-
-	const imagesToDisplay = isString(props.modalOpenedBy)
-		? exercise?.images || []
-		: currentWorkout.images;
-
-	const videosToDisplay = isString(props.modalOpenedBy)
-		? exercise?.videos || []
-		: currentWorkout.videos;
-
-	const hasAssets = imagesToDisplay.length > 0 || videosToDisplay.length > 0;
-
-	const onRemoveAsset = (key: string, type: "image" | "video") => {
-		deleteUploadedAsset(key);
-		setCurrentWorkout(
-			produce(currentWorkout, (draft) => {
-				if (type === "image") {
-					if (exerciseIdx !== -1) {
-						draft.exercises[exerciseIdx].images = draft.exercises[
-							exerciseIdx
-						].images.filter((i) => i !== key);
-					} else {
-						draft.images = draft.images.filter((i) => i !== key);
-					}
-					return;
-				}
-				if (exerciseIdx !== -1) {
-					draft.exercises[exerciseIdx].videos = draft.exercises[
-						exerciseIdx
-					].videos.filter((i) => i !== key);
-				} else {
-					draft.videos = draft.videos.filter((i) => i !== key);
-				}
-			}),
-		);
-	};
-
-	return (
-		<Modal
-			onClose={() => props.closeModal()}
-			opened={props.modalOpenedBy !== undefined}
-			title={`Images for ${exerciseDetails ? exerciseDetails.name : "the workout"}`}
-		>
-			<Stack>
-				{fileUploadAllowed ? (
-					<>
-						{hasAssets ? (
-							<Avatar.Group spacing="xs">
-								{imagesToDisplay.map((i) => (
-									<AssetDisplay
-										key={i}
-										s3Key={i}
-										type="image"
-										removeAsset={() => onRemoveAsset(i, "image")}
-									/>
-								))}
-								{videosToDisplay.map((i) => (
-									<AssetDisplay
-										key={i}
-										s3Key={i}
-										type="video"
-										removeAsset={() => onRemoveAsset(i, "video")}
-									/>
-								))}
-							</Avatar.Group>
-						) : null}
-						<Group justify="space-between">
-							<FileButton
-								accept="image/*"
-								onChange={(file) => afterFileSelected(file, "image")}
-							>
-								{(props) => (
-									<Button
-										{...props}
-										flex={1}
-										color="cyan"
-										variant="outline"
-										loading={isFileUploading}
-										leftSection={<IconLibraryPhoto />}
-									>
-										Image
-									</Button>
-								)}
-							</FileButton>
-							<FileButton
-								accept="video/*"
-								onChange={(file) => afterFileSelected(file, "video")}
-							>
-								{(props) => (
-									<Button
-										{...props}
-										flex={1}
-										color="cyan"
-										variant="outline"
-										loading={isFileUploading}
-										leftSection={<IconVideo />}
-									>
-										Video
-									</Button>
-								)}
-							</FileButton>
-						</Group>
-					</>
-				) : (
-					<Text c="red" size="sm">
-						Please set the S3 variables required to enable file uploading
-					</Text>
-				)}
-			</Stack>
-		</Modal>
-	);
-};
